chore: revert page background support
