+++ packages/widget-playground/src/components/DrawerControls/DesignControls/LayoutControls.tsx
@@ -2,7 +2,6 @@ import { defaultMaxHeight } from '@lifi/widget';
 import { MenuItem, type SelectChangeEvent } from '@mui/material';
 import type { CSSProperties } from 'react';
 import { type ChangeEventHandler, useEffect, useId, useState } from 'react';
-import { useDefaultViewportColor } from '../../../hooks';
 import {
   type Layout,
   useConfig,
@@ -11,7 +10,6 @@ import {
   useEditToolsActions,
   useHeaderAndFooterToolValues,
   useLayoutValues,
-  usePlaygroundSettingValues,
 } from '../../../store';
 import {
   CardRowColumn,
@@ -84,8 +82,6 @@ const getLayoutMode = (container?: CSSProperties) => {
 export const LayoutControls = () => {
   const inputId = useId();
   const { config } = useConfig();
-  const { viewportColor } = usePlaygroundSettingValues();
-  const { defaultColor } = useDefaultViewportColor();
 
   const { variant } = useConfigVariant();
   const { showMockHeader } = useHeaderAndFooterToolValues();
@@ -140,15 +136,6 @@ export const LayoutControls = () => {
         setHeader({
           position: 'fixed',
           top: showMockHeader ? 48 : 0,
-          pageBackground: viewportColor ?? defaultColor,
-          ...(getCurrentConfigTheme()?.container?.borderRadius
-            ? {
-                borderTopLeftRadius:
-                  getCurrentConfigTheme()?.container?.borderRadius,
-                borderTopRightRadius:
-                  getCurrentConfigTheme()?.container?.borderRadius,
-              }
-            : {}),
         });
 
         const fullHeightContainer = {
