+++ packages/widget-playground/src/components/DrawerControls/DesignControls/PlaygroundSettingsControl/ViewportColorSelector.tsx
@@ -1,8 +1,8 @@
 import type { BoxProps } from '@mui/material';
-import { useDefaultViewportColor } from '../../../../hooks';
+import { useTheme } from '@mui/material';
 import {
+  useEditToolsActions,
   usePlaygroundSettingValues,
-  useSetViewportBackgroundColor,
 } from '../../../../store';
 import { safe6DigitHexColor } from '../../../../utils';
 import {
@@ -12,9 +12,14 @@ import {
 } from '../DesignControls.style';
 
 export const ViewportColorSelector = ({ ...rest }: BoxProps) => {
-  const { defaultColor } = useDefaultViewportColor();
+  const theme = useTheme();
   const { viewportColor } = usePlaygroundSettingValues();
-  const { setViewportBackgroundColor } = useSetViewportBackgroundColor();
+  const { setViewportBackgroundColor } = useEditToolsActions();
+
+  const defaultColor =
+    theme.palette.mode === 'light'
+      ? theme.palette.grey[100]
+      : theme.palette.common.black;
 
   return (
     <ColorControlContainer {...rest}>
