+++ packages/widget-playground/src/hooks/index.ts
@@ -1,3 +1,2 @@
-export * from './useDefaultViewportColor';
 export * from './useDevView';
 export * from './useThemeMode';
