+++ packages/widget-playground/src/store/widgetConfig/useConfig.ts
@@ -8,14 +8,3 @@ export const useConfig = () => {
     config,
   };
 };
-
-export const useThemeHeader = () => {
-  const header = useWidgetConfigStore(
-    (state) => state.config?.theme?.header,
-    shallow,
-  );
-
-  return {
-    header,
-  };
-};
